refactor(ui): extract api/coordinator-admin.ts — all coordinator/admin endpoints
Move the 15 coordinator-admin viewer endpoints into
api/coordinator-admin.ts: status load, group CRUD (create, rename,
archive, unarchive, filtered list), join-request review, device
lifecycle (rename, disable, enable, remove), and invite creation.
The barrel re-exports each function so the coordinator-admin tab
actions keep the same import path.
